+++ src/store/validators/actions.js
@@ -1,4 +1,5 @@
-import { staking, validators } from '@/apis/http';
+import { bank, distribution, staking, validators } from '@/apis/http';
+import msgBuilder from '@/store/keplr/helpers/msgBuilder';
 
 export default {
   async initValidatorsList({ commit, dispatch }) {
@@ -8,7 +9,6 @@ export default {
     await Promise.all(requests);
     commit('setLoading', false);
   },
-
   async fetchList({ commit }) {
     try {
       const response = await validators.requestList();
@@ -17,7 +17,6 @@ export default {
       commit('setError', error);
     }
   },
-
   async initValidatorsDetail({ commit, dispatch }, address) {
     commit('reset');
     commit('setLoading', true);
@@ -28,13 +27,11 @@ export default {
     await Promise.all(requests);
     commit('setLoading', false);
   },
-
   async updateValidatorsDetail({ commit, dispatch }, address) {
     commit('setUpdating', true);
     await dispatch('fetchDetail', address);
     commit('setUpdating', false);
   },
-
   async fetchDetail({ commit }, address) {
     try {
       const response = await validators.requestDetail(address);
@@ -43,7 +40,6 @@ export default {
       commit('setError', error);
     }
   },
-
   async fetchDetailDelegations({ dispatch, getters }, id) {
     await dispatch('addDetailDelegations', { id });
     while (getters['delegationsTotal'] > getters['delegationsOffset']) {
@@ -53,7 +49,6 @@ export default {
       });
     }
   },
-
   async addDetailDelegations({ commit }, { id, offset }) {
     const pagination = {
       offset: offset ? offset : 0,
@@ -70,8 +65,162 @@ export default {
       commit('setError', error);
     }
   },
-
   setValidatorsFilter({ commit }, filter) {
     commit('setFilter', filter);
+  },
+  async initWallet({ commit, dispatch, rootGetters }) {
+    commit('resetWallet');
+    commit('setLoadingWallet', true);
+    const accounts = rootGetters['keplr/accounts'];
+    const account = accounts[0].address;
+    const requests = [
+      dispatch('fetchAccountBalances', account),
+      dispatch('fetchAccountDelegations', account),
+      dispatch('fetchAccountRewards', account),
+      dispatch('fetchAccountUnbondings', account),
+    ];
+    await Promise.all(requests);
+    commit('setLoadingWallet', false);
+  },
+  async fetchAccountBalances({ commit }, account) {
+    try {
+      const response = await bank.requestBalances(account);
+      commit('setWalletItem', {
+        balances: response.data.balances,
+      });
+    } catch {
+      commit('setWalletItem', { balances: [] });
+      return;
+    }
+  },
+  async fetchAccountDelegations({ commit }, account) {
+    try {
+      const response = await staking.requestDelegations(account);
+      commit('setWalletItem', {
+        delegations: response.data.delegation_responses,
+      });
+    } catch {
+      commit('setWalletItem', { delegations: [] });
+      return;
+    }
+  },
+  async fetchAccountUnbondings({ dispatch, getters }, account) {
+    await dispatch('getWalletUnbondings', { account });
+    while (
+      getters['walletUnbondingsTotal'] > getters['walletUnbondingsOffset']
+    ) {
+      await dispatch('getWalletUnbondings', {
+        account,
+        offset: getters['walletUnbondingsOffset'],
+      });
+    }
+  },
+  async getWalletUnbondings({ commit }, { account, offset }) {
+    const pagination = {
+      offset: offset ? offset : 0,
+    };
+    try {
+      const response = await staking.requestUnbondings(account, pagination);
+      commit('addWalletUnbondings', response.data.unbonding_responses);
+      commit('setWalletUnbondingsPagination', response.data.pagination);
+      commit(
+        'sumWalletUnbondingsOffset',
+        response.data.unbonding_responses.length
+      );
+    } catch (error) {
+      commit('setWalletItem', { unbondings: [] });
+      return;
+    }
+  },
+  async fetchAccountRewards({ commit }, account) {
+    try {
+      const response = await distribution.requestRewards(account);
+      commit('setWalletItem', {
+        rewards: response.data,
+      });
+    } catch (error) {
+      commit('setWalletItem', { rewards: [] });
+      return;
+    }
+  },
+  async delegateTokens(
+    { commit, dispatch, rootGetters },
+    { validatorAddress, amount, translator, context }
+  ) {
+    commit('setLoading', true);
+    const accounts = rootGetters['keplr/accounts'];
+    const account = accounts[0].address;
+    const msg = msgBuilder.buildMsgDelegate({
+      validatorAddress,
+      amount,
+      account,
+    });
+    await dispatch(
+      'keplr/signAndBroadcastTransaction',
+      { msgs: [msg], translator, context },
+      { root: true }
+    );
+    commit('setLoading', false);
+  },
+  async undelegateTokens(
+    { commit, dispatch, rootGetters },
+    { validatorAddress, amount, translator, context }
+  ) {
+    commit('setLoading', true);
+    const accounts = rootGetters['keplr/accounts'];
+    const account = accounts[0].address;
+    const msg = msgBuilder.buildMsgUndelegate({
+      validatorAddress,
+      amount,
+      account,
+    });
+    await dispatch(
+      'keplr/signAndBroadcastTransaction',
+      { msgs: [msg], translator, context },
+      { root: true }
+    );
+    commit('setLoading', false);
+  },
+  async redelagateTokens(
+    { commit, dispatch, rootGetters },
+    { validatorAddress, srcAddress, amount, translator, context }
+  ) {
+    commit('setLoading', true);
+    const accounts = rootGetters['keplr/accounts'];
+    const account = accounts[0].address;
+    const msg = msgBuilder.buildMsgBeginRedelegate({
+      validatorAddress,
+      srcAddress,
+      amount,
+      account,
+    });
+    await dispatch(
+      'keplr/signAndBroadcastTransaction',
+      { msgs: [msg], translator, context },
+      { root: true }
+    );
+    commit('setLoading', false);
+  },
+  async claimRewards(
+    { commit, dispatch, rootGetters },
+    { validators, translator, context }
+  ) {
+    commit('setLoading', true);
+    const accounts = rootGetters['keplr/accounts'];
+    const account = accounts[0].address;
+    const msgs = [];
+    for (const validatorAddress of validators) {
+      const msg = msgBuilder.buildMsgWithdrawDelegatorReward({
+        validatorAddress,
+        account,
+      });
+      msgs.push(msg);
+    }
+    await dispatch(
+      'keplr/signAndBroadcastTransaction',
+      { msgs: msgs, translator, context },
+      { root: true }
+    );
+    commit('setLoading', false);
   },
 };