+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.3.0] - 2022-10-14
+
+### Added
+
+- Add transaction execution functions in the chain using Keplr:  
+  1. Delegating tokens
+  2. Undelegating tokens
+  3. Redelegating tokens
+  4. Claim tokens
+  5. Voting on proposals
+- Add local wallet update when changed by Keplr
+- Add proposal stats
+
+### Changed
+
+- Update value of voting transactions filter
+
 ## [4.2.3] - 2022-09-21
 
 ### Changed
@@ -497,6 +514,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add RFC cross chain navigation
 - Add a switch to selcet a dark theme from navigation drawer
 
+[4.3.0]: https://github.com/commercionetwork/almerico/compare/v4.2.3...v4.3.0
 [4.2.3]: https://github.com/commercionetwork/almerico/compare/v4.2.2...v4.2.3
 [4.2.2]: https://github.com/commercionetwork/almerico/compare/v4.2.1...v4.2.2
 [4.2.1]: https://github.com/commercionetwork/almerico/compare/v4.2.0...v4.2.1